requirements.txt
Pillow==8.4.0
mysql-connector-python==8.0.27
